delete old resolvers from apiUtils package.
