+++ packages/apiUtils/src/index.ts
@@ -1,5 +1,4 @@
 export * from './config';
-export * from './utils';
 export * from './resolvers';
 export * from './resolvers2';
 export * from './ApiUtils';
