--- packages/apiUtils/src/resolvers/Endpoints.ts
@@ -1,69 +1,0 @@
-import Core from '@moralisweb3/common-core';
-import { Endpoint, EndpointFactory, EndpointMethod } from './Endpoint';
-import { EndpointResolver } from './EndpointResolver';
-import { PaginatedEndpointFactory, PaginatedParams } from './PaginatedEndpoint';
-import { PaginatedEndpointResolver } from './PaginatedEndpointResolver';
-
-// TODO: this interface could be replaced by Endpoint<> interface, but this is not possible for now until
-// changes described in this file are not implemented.
-export interface EndpointDescriptor {
-  name: string;
-  urlPatternParamNames: string[];
-  urlPattern: string;
-  bodyParamNames: string[];
-  method: EndpointMethod;
-}
-
-export class Endpoints {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  private readonly endpoints: Endpoint<any, any, any, any, unknown>[] = [];
-
-  public constructor(private readonly core: Core, private readonly baseUrl: string) {}
-
-  public createFetcher<ApiParams, Params, ApiResult, AdaptedResult, JSONResult>(
-    factory: EndpointFactory<ApiParams, Params, ApiResult, AdaptedResult, JSONResult>,
-  ) {
-    const resolver = EndpointResolver.create(this.core, this.baseUrl, factory);
-    this.endpoints.push(resolver.endpoint);
-    return resolver.fetch;
-  }
-
-  public createNullableFetcher<ApiParams, Params, ApiResult, AdaptedResult, JSONResult>(
-    factory: EndpointFactory<ApiParams, Params, ApiResult, AdaptedResult, JSONResult>,
-  ) {
-    const resolver = EndpointResolver.create(this.core, this.baseUrl, factory);
-    this.endpoints.push(resolver.endpoint);
-    return resolver.fetchNullable;
-  }
-
-  public createPaginatedFetcher<ApiParams, Params extends PaginatedParams, ApiResult, AdaptedResult, JSONResult>(
-    factory: PaginatedEndpointFactory<ApiParams, Params, ApiResult, AdaptedResult, JSONResult>,
-  ) {
-    const resolver = PaginatedEndpointResolver.create(this.core, this.baseUrl, factory);
-    this.endpoints.push(resolver.endpoint);
-    return resolver.fetch;
-  }
-
-  public getDescriptors(): EndpointDescriptor[] {
-    return this.endpoints.map((endpoint) => {
-      const urlPatternParamNames = (endpoint.urlParams as string[]) || [];
-
-      // TODO: Endpoint<> type should have `urlPattern` property instead of `getUrl` method!
-      // For example: .urlPattern = '/{token0Address}/{token1Address}/pairAddress'
-      const urlParams = urlPatternParamNames.reduce<{ [name: string]: string }>((params, paramName) => {
-        params[paramName] = `{${paramName}}`;
-        return params;
-      }, {});
-      const urlPattern = endpoint.getUrl(urlParams);
-
-      return {
-        // DO NOT return baseUrl here!
-        name: endpoint.name,
-        urlPatternParamNames,
-        urlPattern,
-        method: endpoint.method || 'get',
-        bodyParamNames: endpoint.bodyParams ? endpoint.bodyParams.map(String) : [],
-      };
-    });
-  }
-}
