+++ packages/apiUtils/src/resolvers/index.ts
@@ -1,8 +1,2 @@
-export * from './Endpoint';
-export * from './Endpoints';
-export * from './EndpointResolver';
-export * from './PaginatedEndpoint';
-export * from './PaginatedEndpointResolver';
-export * from './ApiResultAdapter';
-export * from './ApiPaginatedResultAdapter';
+export * from './EndpointDescriptor';
 export * from './DescriptorConverter';
