+++ packages/apiUtils/src/resolvers2/getCommonHeaders.ts
