--- packages/core/src/version.ts
@@ -1,1 +1,0 @@
-export const LIB_VERSION = '2.7.0';
